Fix incorrect "this" binding when there's a TypeScript `this` arg in scope
Our `mayUseLexicalThis` guard was not fully guarding against treating a template's `this` as the lexically scoped one. This went unnoticed because to trigger the bug you also need babel to tell you that the identifier `this` is in scope, which normally won't happen, except in TypeScript with a `this` function argument.