Don't accidentally load ambient config during bindExpression
`babel.parse` will search for a config file in the CWD and use it. This will explode if the config file contains async features, since we're doing a synchronous parse.

This could be fixed by passing `configFile: false` to parse, but I also realized I could use `babel.template` instead and that is simpler anyway.